Overhaul Enum/StringRegistry infrastructure: zero-heap registration, constexpr IDs, NTTP migration
Replace all phantom struct tag types with CompiledString non-type template
parameters (NTTPs) throughout VariantDatabase and its subclasses
(MediaConfig, Metadata, LibraryOptions, MediaIOStats, MediaIOParams, Config,
ClockDomain, Benchmark).  This eliminates seven phantom tag structs and makes
each database's name a true string literal visible in diagnostics.

Introduce PROMEKI_REGISTER_ENUM_TYPE macro that pins an Enum entry table to
rodata via a static constexpr array, with three compile-time static_asserts
(non-empty table, unique names/values, default present in table).  Add
Enum::registerDefinition() as a zero-heap registration path alongside the
existing heap-allocating registerType().  Migrate all well-known enums in
enums.h to the new macro.

Add literal-backed String caches to Enum::Definition (nameLit,
entryNameLits, entryQualifiedLits) so that typeName(), valueName(), and
in-list toString() return zero-copy StringLiteralData-backed Strings with no
allocation on the hot path.  Two Enum values sharing the same in-list value
now return Strings backed by the same cstr() pointer, allowing String ==
to short-circuit on pointer identity.

Introduce PROMEKI_DECLARE_ID macro to replace the static inline const ID =
declareID(...) pattern so that well-known IDs declared in class bodies are
constexpr-compatible (usable in switch labels, static_assert, template
parameters).

Add PROMEKI_INTENTIONAL_LEAK(ptr) macro to util.h: delegates to
__lsan_ignore_object() under ASAN/LSAN, compiles to a no-op otherwise.
Annotate all three intentional-leak sites in enum.cpp.  Add
etc/valgrind.supp with matching Valgrind suppressions.

Make CompiledString data members public to satisfy the C++20 structural type
requirement for class-type NTTPs.

Delete stringregistry_constexpr.md (work complete).  Update
CODING_STANDARDS.md well-known enum example to use the new macro.  Expand
devplan/README.md Phase 4f paragraph.  Add tests covering all new paths.

+++ src/core/clockdomain.cpp
@@ -28,7 +28,7 @@ struct ClockDomain::Data {
 // ============================================================================
 
 struct ClockDomainRegistry {
-        Map<uint32_t, ClockDomain::Data> entries;
+        Map<uint64_t, ClockDomain::Data> entries;
 
         ClockDomainRegistry() {
                 // Synthetic — frame-rate derived, per-stream epoch
@@ -52,7 +52,7 @@ struct ClockDomainRegistry {
         }
 
         void add(ClockDomain::Data &&d) {
-                uint32_t key = d.id.id();
+                uint64_t key = d.id.id();
                 entries.insert(key, std::move(d));
         }
